Refactor: nonce, sha256 로직 뷰모델로 분리
