+++ Kabinett/Presentation/View/Profile/LoginView.swift
@@ -10,7 +10,7 @@ import AuthenticationServices
 import CryptoKit
 
 struct LoginView: View {
-    @State private var currentNonce: String?
+    @StateObject private var viewModel = SignUpViewModel(useCase: SignUpUseCaseStub())
     
     var body: some View {
         GeometryReader { geometry in
@@ -36,18 +36,18 @@ struct LoginView: View {
                         .padding(.bottom, 25)
                     
                     SignInWithAppleButton { request in
-                        handleSignInWithAppleRequest(request)
+                        viewModel.handleSignInWithAppleRequest(request)
                     } onCompletion: { result in
                         switch result {
                         case .success(let authorization):
                             if let appleIDCredential = authorization.credential as? ASAuthorizationAppleIDCredential {
-                                    let userIdentifier = appleIDCredential.user
-                                    let userEmail = appleIDCredential.email ?? "이메일 정보 없음"
-                                    print("사용자 ID: \(userIdentifier)")
-                                    print("이메일: \(userEmail)")
-                                } else {
-                                    print("Apple ID credential 변환에 실패했습니다.")
-                                }
+                                let userIdentifier = appleIDCredential.user
+                                let userEmail = appleIDCredential.email ?? "이메일 정보 없음"
+                                print("사용자 ID: \(userIdentifier)")
+                                print("이메일: \(userEmail)")
+                            } else {
+                                print("Apple ID credential 변환에 실패했습니다.")
+                            }
                         case .failure(let error):
                             print("Authorization failed: \(error.localizedDescription)")
                         }
@@ -60,57 +60,8 @@ struct LoginView: View {
             }
         }
     }
-    
-    private func handleSignInWithAppleRequest(_ request: ASAuthorizationAppleIDRequest) {
-        request.requestedScopes = [.email]
-        let nonce = randomNonceString()
-        currentNonce = nonce
-        request.nonce = sha256(nonce)
-    }
-    
-    private func randomNonceString(length: Int = 32) -> String {
-        precondition(length > 0)
-        let charset: [Character] =
-        Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
-        var result = ""
-        var remainingLength = length
-        
-        while remainingLength > 0 {
-            let randoms: [UInt8] = (0 ..< 16).map { _ in
-                var random: UInt8 = 0
-                let errorCode = SecRandomCopyBytes(kSecRandomDefault, 1, &random)
-                if errorCode != errSecSuccess {
-                    fatalError("Unable to generate nonce. SecRandomCopyBytes failed with OSStatus \(errorCode).")
-                }
-                return random
-            }
-            
-            randoms.forEach { random in
-                if remainingLength == 0 {
-                    return
-                }
-                
-                if random < charset.count {
-                    result.append(charset[Int(random)])
-                    remainingLength -= 1
-                }
-            }
-        }
-        
-        return result
-    }
-    
-    private func sha256(_ input: String) -> String {
-        let inputData = Data(input.utf8)
-        let hashedData = SHA256.hash(data: inputData)
-        let hashString = hashedData.compactMap {
-            String(format: "%02x", $0)
-        }.joined()
-        
-        return hashString
-    }
 }
-
+    
 
 #Preview {
     LoginView()
